Refactor UI directory structure and modernize MapWindow event handling
- Moved `*_window.cpp/h` to `source/ui/windows/`.
- Moved `*_dialog.cpp/h` to `source/ui/dialogs/`.
- Updated `source/CMakeLists.txt` with new file paths.
- Updated include paths in all affected files.
- Removed legacy `BEGIN_EVENT_TABLE` macros from `source/app/application.cpp`.
- Removed `DECLARE_EVENT_TABLE` from `MapWindow` and `MapScrollBar`.
- Implemented `Bind` for `MapScrollBar` events in its constructor.
- Added missing `wxEVT_SCROLL_TOP` and `wxEVT_SCROLL_BOTTOM` bindings to `MapWindow`.
