+++ source/app/application.cpp
@@ -27,9 +27,9 @@
 #include "ui/dialogs/goto_position_dialog.h"
 #include "palette/palette_window.h"
 #include "app/preferences.h"
-#include "ui/result_window.h"
+#include "ui/windows/result_window.h"
 #include "rendering/ui/minimap_window.h"
-#include "ui/about_window.h"
+#include "ui/windows/about_window.h"
 #include "ui/main_menubar.h"
 #include "app/updater.h"
 #include "ui/artprovider.h"
@@ -46,36 +46,6 @@
 #include <spdlog/spdlog.h>
 
 #include "../brushes/icon/editor_icon.xpm"
-
-BEGIN_EVENT_TABLE(MapWindow, wxPanel)
-EVT_SIZE(MapWindow::OnSize)
-
-EVT_COMMAND_SCROLL_TOP(MAP_WINDOW_HSCROLL, MapWindow::OnScroll)
-EVT_COMMAND_SCROLL_BOTTOM(MAP_WINDOW_HSCROLL, MapWindow::OnScroll)
-EVT_COMMAND_SCROLL_THUMBTRACK(MAP_WINDOW_HSCROLL, MapWindow::OnScroll)
-EVT_COMMAND_SCROLL_LINEUP(MAP_WINDOW_HSCROLL, MapWindow::OnScrollLineUp)
-EVT_COMMAND_SCROLL_LINEDOWN(MAP_WINDOW_HSCROLL, MapWindow::OnScrollLineDown)
-EVT_COMMAND_SCROLL_PAGEUP(MAP_WINDOW_HSCROLL, MapWindow::OnScrollPageUp)
-EVT_COMMAND_SCROLL_PAGEDOWN(MAP_WINDOW_HSCROLL, MapWindow::OnScrollPageDown)
-
-EVT_COMMAND_SCROLL_TOP(MAP_WINDOW_VSCROLL, MapWindow::OnScroll)
-EVT_COMMAND_SCROLL_BOTTOM(MAP_WINDOW_VSCROLL, MapWindow::OnScroll)
-EVT_COMMAND_SCROLL_THUMBTRACK(MAP_WINDOW_VSCROLL, MapWindow::OnScroll)
-EVT_COMMAND_SCROLL_LINEUP(MAP_WINDOW_VSCROLL, MapWindow::OnScrollLineUp)
-EVT_COMMAND_SCROLL_LINEDOWN(MAP_WINDOW_VSCROLL, MapWindow::OnScrollLineDown)
-EVT_COMMAND_SCROLL_PAGEUP(MAP_WINDOW_VSCROLL, MapWindow::OnScrollPageUp)
-EVT_COMMAND_SCROLL_PAGEDOWN(MAP_WINDOW_VSCROLL, MapWindow::OnScrollPageDown)
-
-EVT_BUTTON(MAP_WINDOW_GEM, MapWindow::OnGem)
-END_EVENT_TABLE()
-
-BEGIN_EVENT_TABLE(MapScrollBar, wxScrollBar)
-EVT_KEY_DOWN(MapScrollBar::OnKey)
-EVT_KEY_UP(MapScrollBar::OnKey)
-EVT_CHAR(MapScrollBar::OnKey)
-EVT_SET_FOCUS(MapScrollBar::OnFocus)
-EVT_MOUSEWHEEL(MapScrollBar::OnWheel)
-END_EVENT_TABLE()
 
 wxIMPLEMENT_APP(Application);
