+++ source/app/application.h
@@ -25,7 +25,7 @@
 
 #include "net/process_com.h"
 #include "rendering/ui/map_display.h"
-#include "ui/welcome_dialog.h"
+#include "ui/dialogs/welcome_dialog.h"
 
 #include "ui/main_frame.h"
