+++ source/rendering/ui/keyboard_handler.cpp
@@ -20,7 +20,7 @@
 #include "rendering/ui/zoom_controller.h"
 #include "rendering/ui/navigation_controller.h"
 #include "rendering/ui/map_display.h"
-#include "ui/map_window.h"
+#include "ui/windows/map_window.h"
 #include "ui/gui.h"
 #include "editor/hotkey_manager.h"
 #include "editor/editor.h"
