+++ source/ui/main_frame.cpp
@@ -15,9 +15,9 @@
 #include "ui/dialogs/goto_position_dialog.h"
 #include "palette/palette_window.h"
 #include "app/preferences.h"
-#include "ui/result_window.h"
+#include "ui/windows/result_window.h"
 #include "rendering/ui/minimap_window.h"
-#include "ui/about_window.h"
+#include "ui/windows/about_window.h"
 #include "ui/main_menubar.h"
 #include "app/updater.h"
 #include "ui/artprovider.h"
