+++ source/ui/windows/find_item_window.cpp
@@ -16,7 +16,7 @@
 //////////////////////////////////////////////////////////////////////
 
 #include "app/main.h"
-#include "ui/find_item_window.h"
+#include "ui/windows/find_item_window.h"
 #include "ui/dialogs/find_dialog.h"
 #include "ui/controls/sortable_list_box.h"
 #include "ui/gui.h"
