+++ source/ui/windows/map_window.h
@@ -19,7 +19,7 @@
 #define RME_MAP_WINDOW_H_
 
 #include "map/position.h"
-#include "ui/replace_items_window.h"
+#include "ui/windows/replace_items_window.h"
 
 class MapCanvas;
 class DCButton;
@@ -97,8 +97,6 @@ private:
 
 	friend class MainFrame;
 	friend class MapCanvas;
-
-	DECLARE_EVENT_TABLE()
 };
 
 // MapScrollbar, a special scrollbar that relays alot of events
@@ -107,7 +105,13 @@ private:
 class MapScrollBar : public wxScrollBar {
 public:
 	MapScrollBar(MapWindow* parent, wxWindowID id, long style, wxWindow* canvas) :
-		wxScrollBar(parent, id, wxDefaultPosition, wxDefaultSize, style), canvas(canvas) { }
+		wxScrollBar(parent, id, wxDefaultPosition, wxDefaultSize, style), canvas(canvas) {
+		Bind(wxEVT_KEY_DOWN, &MapScrollBar::OnKey, this);
+		Bind(wxEVT_KEY_UP, &MapScrollBar::OnKey, this);
+		Bind(wxEVT_CHAR, &MapScrollBar::OnKey, this);
+		Bind(wxEVT_SET_FOCUS, &MapScrollBar::OnFocus, this);
+		Bind(wxEVT_MOUSEWHEEL, &MapScrollBar::OnWheel, this);
+	}
 	virtual ~MapScrollBar() { }
 
 	void OnKey(wxKeyEvent& event) {
@@ -121,7 +125,6 @@ public:
 	}
 
 	wxWindow* canvas;
-	DECLARE_EVENT_TABLE()
 };
 
 #endif
